test for empty show
